Revert "Turbinia API processor modules (#747)"
This reverts commit 41d49877a4d0f5b98b104404de62287503faaccf.
