+++ data/config.json
@@ -1,5 +1,5 @@
 {
-  "ts_username": "timesketch",
-  "ts_password": "timesketch",
-  "ts_endpoint": "http://127.0.0.1:5000"
+  "ts_username": "your_username",
+  "ts_password": "your_password",
+  "ts_endpoint": "http://127.0.0.1:80"
 }
