+++ data/recipes/aws_turbinia_ts.json
@@ -63,14 +63,11 @@
       "wants": ["GCEDiskFromImage"],
       "name": "TurbiniaGCPProcessor",
       "args": {
+        "turbinia_config_file": null,
         "disk_names": null,
-        "request_ids": "@request_ids",
-        "project": "@analysis_project_name",
-        "turbinia_auth": "@turbinia_auth",
+        "project": "@gcp_project",
         "turbinia_recipe": "@turbinia_recipe",
         "turbinia_zone": "@turbinia_zone",
-        "turbinia_api": "@turbinia_api",
-        "incident_id": "@incident_id",
         "sketch_id": "@sketch_id"
       }
     }, {
@@ -97,11 +94,8 @@
     ["--incident_id", "Incident ID (used for Timesketch description).", null],
     ["--sketch_id", "Timesketch sketch to which the timeline should be added.", null, {"format": "regex", "regex": "^\\d+$"}],
     ["--token_password", "Optional custom password to decrypt Timesketch credential file with.", ""],
-    ["--request_ids", "Comma separated Turbinia request identifiers to process.", null, {"format": "regex", "comma_separated": true, "regex": "^[a-f0-9]{32}$"}],
     ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
     ["--turbinia_zone", "Zone Turbinia is located in", "us-central1-f", {"format": "gcp_zone"}],
-    ["--turbinia_auth", "Flag to indicate whether Turbinia API server requires authentication.", false],
-    ["--turbinia_api", "Turbinia API server endpoint.", "http://127.0.0.1:8000"],
     ["--wait_for_timelines", "Whether to wait for Timesketch to finish processing all timelines.", true]
   ]
 }
