--- data/recipes/aws_turbinia_ts_legacy.json
@@ -1,101 +1,0 @@
-{
-  "name": "aws_turbinia_ts_legacy",
-  "short_description": "Copies EBS volumes from within AWS, transfers them to GCP, analyses with Turbinia and exports the results to Timesketch.",
-  "description": "Copies EBS volumes from within AWS, uses buckets and cloud-to-cloud operations to transfer the data to GCP. Once in GCP, a persistent disk is created and a job is added to the Turbinia queue to start analysis. The resulting Plaso file is then exported to Timesketch.",
-  "preflights": [
-    {
-      "wants": [],
-      "name": "AWSAccountCheck",
-      "args": {
-        "profile_name": "@aws_profile"
-      }
-    },{
-      "wants": [],
-      "name": "GCPTokenCheck",
-      "args": {
-        "project_name": "@gcp_project"
-      }
-    }
-  ],
-  "modules": [
-    {
-      "wants": [],
-      "name": "AWSVolumeSnapshotCollector",
-      "args": {
-        "volumes": "@volumes",
-        "region": "@aws_region"
-      }
-    },{
-      "wants": ["AWSVolumeSnapshotCollector"],
-      "name": "AWSSnapshotS3CopyCollector",
-      "args": {
-        "snapshots": "",
-        "bucket": "@aws_bucket",
-        "region": "@aws_region",
-        "subnet": "@subnet"
-      }
-    },{
-      "wants": ["AWSSnapshotS3CopyCollector"],
-      "name": "S3ToGCSCopy",
-      "args": {
-        "s3_objects": "",
-        "aws_region": "@aws_region",
-        "dest_project": "@gcp_project",
-        "dest_bucket": "@gcp_bucket",
-        "object_filter": ".+/image.bin$"
-      }
-    }, {
-      "wants": ["S3ToGCSCopy"],
-      "name": "GCSToGCEImage",
-      "args": {
-        "source_objects": "",
-        "dest_project": "@gcp_project"
-      }
-    }, {
-      "wants": ["GCSToGCEImage"],
-      "name": "GCEDiskFromImage",
-      "args": {
-        "source_images": "",
-        "dest_project": "@gcp_project",
-        "dest_zone": "@gcp_zone"
-      }
-    }, {
-      "wants": ["GCEDiskFromImage"],
-      "name": "TurbiniaGCPProcessorLegacy",
-      "args": {
-        "turbinia_config_file": null,
-        "disk_names": null,
-        "project": "@gcp_project",
-        "turbinia_recipe": "@turbinia_recipe",
-        "turbinia_zone": "@turbinia_zone",
-        "sketch_id": "@sketch_id"
-      }
-    }, {
-      "wants": ["TurbiniaGCPProcessorLegacy"],
-      "name": "TimesketchExporter",
-      "args": {
-        "incident_id": "@incident_id",
-        "token_password": "@token_password",
-        "sketch_id": "@sketch_id",
-        "analyzers": null,
-        "wait_for_timelines": "@wait_for_timelines"
-      }
-    }
-  ],
-  "args": [
-    ["aws_region", "AWS region containing the EBS volumes.", null, {"format": "aws_region"}],
-    ["gcp_zone", "Destination GCP zone in which to create the disks.", null, {"format": "gcp_zone"}],
-    ["volumes", "Comma separated list of EBS volume IDs (e.g. vol-xxxxxxxx).", null, {"format": "regex", "comma_separated": true, "regex": "^vol-[0-9a-f]{8,17}$"}],
-    ["aws_bucket", "AWS bucket for image storage.", null, {"format": "regex", "comma_separated": false, "regex": "^(s3:\/\/)?[0-9a-z][-\\.0-9a-z]{1,61}[0-9a-z]$"}],
-    ["gcp_bucket", "GCP bucket for image storage.", null, {"format": "regex", "comma_separated": false, "regex": "^(gs:\/\/)?[0-9a-z][-\\.0-9a-z_]{1,61}[0-9a-z]$"}],
-    ["--subnet", "AWS subnet to copy instances from, required if there is no default subnet in the volume region.", null, {"format": "regex", "comma_separated": false, "regex": "^subnet-[0-9a-f]{8,17}$"}],
-    ["--gcp_project", "Destination GCP project.", null, {"format": "regex", "comma_separated": false, "regex": "^[a-z][-a-z0-9]{4,28}[a-z0-9]$"}],
-    ["--aws_profile", "Source AWS profile.", null],
-    ["--incident_id", "Incident ID (used for Timesketch description).", null],
-    ["--sketch_id", "Timesketch sketch to which the timeline should be added.", null, {"format": "regex", "regex": "^\\d+$"}],
-    ["--token_password", "Optional custom password to decrypt Timesketch credential file with.", ""],
-    ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
-    ["--turbinia_zone", "Zone Turbinia is located in", "us-central1-f", {"format": "gcp_zone"}],
-    ["--wait_for_timelines", "Whether to wait for Timesketch to finish processing all timelines.", true]
-  ]
-}
