+++ data/recipes/gcp_turbinia_disk_copy_ts.json
@@ -48,14 +48,11 @@
         "wants": ["GCEDiskCopy"],
         "name": "TurbiniaGCPProcessor",
         "args": {
+            "turbinia_config_file": null,
             "disk_names": null,
-            "request_ids": "@request_ids",
             "project": "@analysis_project_name",
-            "turbinia_auth": "@turbinia_auth",
             "turbinia_recipe": "@turbinia_recipe",
             "turbinia_zone": "@turbinia_zone",
-            "turbinia_api": "@turbinia_api",
-            "incident_id": "@incident_id",
             "sketch_id": "@sketch_id"
         }
     }, {
@@ -72,11 +69,8 @@
     "args": [
         ["source_project_name", "Name of the project containing the instance / disks to copy.", null, {"format": "regex", "comma_separated": false, "regex": "^[a-z][-a-z0-9]{4,28}[a-z0-9]$"}],
         ["analysis_project_name", "Name of the project containing the Turbinia instance.", null, {"format": "regex", "comma_separated": false, "regex": "^[a-z][-a-z0-9]{4,28}[a-z0-9]$"}],
-        ["--request_ids", "Comma separated Turbinia request identifiers to process.", null, {"format": "regex", "comma_separated": true, "regex": "^[a-f0-9]{32}$"}],
         ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
         ["--turbinia_zone", "The GCP zone the disk to process and Turbinia workers are in.", "us-central1-f", {"format": "gcp_zone"}],
-        ["--turbinia_auth", "Flag to indicate whether Turbinia API server requires authentication.", false],
-        ["--turbinia_api", "Turbinia API server endpoint.", "http://127.0.0.1:8000"],
         ["--incident_id", "Incident ID (used for Timesketch description and to label the VM with).", null],
         ["--sketch_id", "Timesketch sketch to which the timeline should be added.", null, {"format": "regex", "regex": "^\\d+$"}],
         ["--token_password", "Optional custom password to decrypt Timesketch credential file with.", ""],
