+++ data/recipes/gcp_turbinia_ts.json
@@ -13,14 +13,11 @@
         "wants": [],
         "name": "TurbiniaGCPProcessor",
         "args": {
-            "request_ids": "@request_ids",
+            "turbinia_config_file": null,
             "disk_names": "@disk_names",
             "project": "@analysis_project_name",
-            "turbinia_auth": "@turbinia_auth",
             "turbinia_recipe": "@turbinia_recipe",
             "turbinia_zone": "@turbinia_zone",
-            "turbinia_api": "@turbinia_api",
-            "incident_id": "@incident_id",
             "sketch_id": "@sketch_id"
         }
     }, {
@@ -37,14 +34,11 @@
     "args": [
         ["analysis_project_name", "Name of GCP project the disk exists in.", null, {"format": "regex", "comma_separated": false, "regex": "^[a-z][-a-z0-9]{4,28}[a-z0-9]$"}],
         ["turbinia_zone", "The GCP zone the disk to process (and Turbinia workers) are in.", null, {"format": "gcp_zone"}],
-        ["--disk_names", "Comma separated names of GCP persistent disks to process. This parameter can only be used if --request_ids is not provided.", null, {"format": "regex", "comma_separated": true, "regex": "^[a-z][-a-z0-9]{0,61}[a-z0-9]?$"}],
-        ["--request_ids", "Comma separated Turbinia request identifiers to process. This parameter can only be used if --disk_names is not provided.", null, {"format": "regex", "comma_separated": true, "regex": "^[a-f0-9]{32}$"}],
+        ["disk_names", "Names of GCP persistent disks to process.", null, {"format": "regex", "comma_separated": true, "regex": "^[a-z][-a-z0-9]{0,61}[a-z0-9]?$"}],
         ["--incident_id", "Incident ID (used for Timesketch description).", null],
         ["--sketch_id", "Timesketch sketch to which the timeline should be added.", null, {"format": "regex", "regex": "^\\d+$"}],
         ["--token_password", "Optional custom password to decrypt Timesketch credential file with.", ""],
         ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
-        ["--turbinia_auth", "Flag to indicate whether Turbinia API server requires authentication.", false],
-        ["--turbinia_api", "Turbinia API server endpoint.", "http://127.0.0.1:8000"],
         ["--wait_for_timelines", "Whether to wait for Timesketch to finish processing all timelines.", true]
     ]
 }
