--- data/recipes/gcp_turbinia_ts_legacy.json
@@ -1,44 +1,0 @@
-{
-    "name": "gcp_turbinia_ts_legacy",
-    "short_description": "Processes existing GCP persistent disks with Turbinia project and sends results to Timesketch.",
-    "description": "Process GCP persistent disks with Turbinia and send output to Timesketch.\n\nThis processes disks that are already in the project where Turbinia exists. If you want to copy disks from another project, use the `gcp_turbinia_disk_copy_ts` recipe.",
-    "preflights": [{
-      "wants": [],
-      "name": "GCPTokenCheck",
-      "args": {
-          "project_name": "@analysis_project_name"
-      }
-    }],
-    "modules": [{
-        "wants": [],
-        "name": "TurbiniaGCPProcessorLegacy",
-        "args": {
-            "turbinia_config_file": null,
-            "disk_names": "@disk_names",
-            "project": "@analysis_project_name",
-            "turbinia_recipe": "@turbinia_recipe",
-            "turbinia_zone": "@turbinia_zone",
-            "sketch_id": "@sketch_id"
-        }
-    }, {
-        "wants": ["TurbiniaGCPProcessorLegacy"],
-        "name": "TimesketchExporter",
-        "args": {
-            "incident_id": "@incident_id",
-            "token_password": "@token_password",
-            "sketch_id": "@sketch_id",
-            "analyzers": null,
-            "wait_for_timelines": "@wait_for_timelines"
-        }
-    }],
-    "args": [
-        ["analysis_project_name", "Name of GCP project the disk exists in.", null, {"format": "regex", "comma_separated": false, "regex": "^[a-z][-a-z0-9]{4,28}[a-z0-9]$"}],
-        ["turbinia_zone", "The GCP zone the disk to process (and Turbinia workers) are in.", null, {"format": "gcp_zone"}],
-        ["disk_names", "Names of GCP persistent disks to process.", null, {"format": "regex", "comma_separated": true, "regex": "^[a-z][-a-z0-9]{0,61}[a-z0-9]?$"}],
-        ["--incident_id", "Incident ID (used for Timesketch description).", null],
-        ["--sketch_id", "Timesketch sketch to which the timeline should be added.", null, {"format": "regex", "regex": "^\\d+$"}],
-        ["--token_password", "Optional custom password to decrypt Timesketch credential file with.", ""],
-        ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
-        ["--wait_for_timelines", "Whether to wait for Timesketch to finish processing all timelines.", true]
-    ]
-}
