+++ data/recipes/upload_turbinia.json
@@ -34,24 +34,36 @@
         "wants": ["SCP-Upload"],
         "name": "TurbiniaArtifactProcessor",
         "args": {
+            "turbinia_config_file": "@turbinia_config",
             "project": null,
-            "turbinia_auth": "@turbinia_auth",
             "turbinia_recipe": "@turbinia_recipe",
             "turbinia_zone": "@turbinia_zone",
-            "turbinia_api": "@turbinia_api",
-            "incident_id": "@incident_id",
             "output_directory": null,
             "sketch_id": null
         }
-    }],
+    }, {
+          "wants": ["TurbiniaArtifactProcessor"],
+          "runtime_name": "SCP-Download",
+          "name": "SCPExporter",
+          "args": {
+            "paths": null,
+            "destination": "@local_turbinia_results",
+            "hostname": "@hostname",
+            "user": null,
+            "id_file": null,
+            "extra_ssh_options": [],
+            "check_ssh": true,
+            "direction": "download",
+            "multiplexing": true
+          }
+      }],
     "args": [
         ["files", "Paths to process.", null],
         ["--turbinia_recipe", "The Turbinia recipe name to use for evidence processing.", null],
         ["--destination_turbinia_dir", "Destination path in Turbinia host to write the files to.", null],
         ["--hostname", "Remote host.", null, {"format": "hostname"}],
         ["--directory", "Directory in which to copy and compress files.", null],
-        ["--turbinia_auth", "Flag to indicate whether Turbinia API server requires authentication.", false],
-        ["--turbinia_api", "Turbinia API server endpoint.", "http://127.0.0.1:8000"],
+        ["--turbinia_config", "Turbinia config file to use.", null],
         ["--local_turbinia_results", "Directory where Turbinia results will be downloaded to.", null],
         ["--turbinia_zone", "The GCP zone the disk to process and Turbinia workers are in.", "us-central1-f", {"format": "gcp_zone"}],
         ["--sketch_id", "Timesketch sketch ID.", null, {"format": "regex", "regex": "^\\d+$"}]
