+++ dftimewolf/lib/exporters/timesketch.py
@@ -3,14 +3,11 @@
 Threaded version of existing Timesketch module."""
 
 import time
-import uuid
-from typing import Optional, List, Type, Union, Set
+from typing import Optional, List, Type, Union
 
 from timesketch_import_client import importer
 from timesketch_api_client import sketch as ts_sketch
 from timesketch_api_client import client as ts_client  # pylint: disable=unused-import,line-too-long  # used for typing
-from timesketch_api_client import error as ts_error
-from timesketch_api_client import analyzer as ts_analyzer
 
 from dftimewolf.lib import module, timesketch_utils
 from dftimewolf.lib.containers import containers, interface
@@ -32,13 +29,13 @@ class TimesketchExporter(module.ThreadAwareModule):
     timesketch_api (TimesketchApiClient): Timesketch API client.
   """
 
+
   sketch: ts_sketch.Sketch
 
-  def __init__(
-      self,
-      state: DFTimewolfState,
-      name: Optional[str] = None,
-      critical: bool = False) -> None:
+  def __init__(self,
+               state: DFTimewolfState,
+               name: Optional[str]=None,
+               critical: bool=False) -> None:
     super(TimesketchExporter, self).__init__(
         state, name=name, critical=critical)
     self.incident_id = None  # type: Union[str, None]
@@ -48,16 +45,13 @@ class TimesketchExporter(module.ThreadAwareModule):
     self.wait_for_timelines = False  # type: bool
     self.host_url = None  # type: Union[str, None]
     self.sketch = None  # type: ts_sketch.Sketch
-    self._processed_timelines: Set[int] = set()
 
-  # pylint: disable=arguments-differ
-  def SetUp(
-      self,
-      incident_id: None = None,
-      sketch_id: Optional[int] = 0,
-      analyzers: None = None,
-      token_password: str = '',
-      wait_for_timelines: bool = False) -> None:
+  def SetUp(self,  # pylint: disable=arguments-differ
+            incident_id: None=None,
+            sketch_id: Optional[int]=0,
+            analyzers: None=None,
+            token_password: str='',
+            wait_for_timelines: bool=False) -> None:
     """Setup a connection to a Timesketch server and create a sketch if needed.
 
     Args:
@@ -81,8 +75,7 @@ class TimesketchExporter(module.ThreadAwareModule):
     if not self.timesketch_api:
       self.ModuleError(
           'Unable to get a Timesketch API client, try deleting the files '
-          '~/.timesketchrc and ~/.timesketch.token',
-          critical=True)
+          '~/.timesketchrc and ~/.timesketch.token', critical=True)
     self.incident_id = incident_id
     self.sketch_id = int(sketch_id) if sketch_id else 0
     self.sketch = None
@@ -112,22 +105,7 @@ class TimesketchExporter(module.ThreadAwareModule):
     if analyzers:
       self._analyzers = [x.strip() for x in analyzers.split(',')]
 
-    self.sketch = self.state.GetFromCache('timesketch_sketch')
-    if not self.sketch and self.sketch_id:
-      self.logger.info('Using exiting sketch: {0:d}'.format(self.sketch_id))
-      self.sketch = self.timesketch_api.get_sketch(self.sketch_id)
-
-    # Create the sketch if no sketch was stored in the cache.
-    if not self.sketch:
-      self.sketch = self._CreateSketch(incident_id=self.incident_id)
-      self.sketch_id = self.sketch.id
-      self.logger.info('New sketch created: {0:d}'.format(self.sketch_id))
-
-    # register callback in timesketch module
-    self.state.RegisterStreamingCallback(self.Process, containers.File)
-
-  def _CreateSketch(
-      self, incident_id: Optional[str] = None) -> ts_sketch.Sketch:
+  def _CreateSketch(self, incident_id: Optional[str]=None) -> ts_sketch.Sketch:
     """Creates a new Timesketch sketch.
 
     Args:
@@ -142,101 +120,84 @@ class TimesketchExporter(module.ThreadAwareModule):
       sketch_name = 'Untitled sketch'
     sketch_description = 'Sketch generated by dfTimewolf'
 
-    sketch = self.timesketch_api.create_sketch(sketch_name, sketch_description)
+    sketch = self.timesketch_api.create_sketch(
+        sketch_name, sketch_description)
     self.sketch_id = sketch.id
     if incident_id:
-      sketch.add_attribute('incident_id', incident_id, ontology='text')
+      sketch.add_attribute(
+          'incident_id', incident_id, ontology='text')
     self.state.AddToCache('timesketch_sketch', sketch)
 
     return sketch
 
   def _WaitForTimelines(self) -> None:
-    """Waits for all timelines in a sketch to be processed.
-
-    Runs analyzers on timelines that are ready.
-    """
+    """Waits for all timelines in a sketch to be processed."""
+    time.sleep(5)  # Give Timesketch time to populate recently added timelines.
     sketch = self.timesketch_api.get_sketch(self.sketch_id)
     timelines = sketch.list_timelines()
-    self.logger.info(
-        f'Found {len(timelines)} timelines for sketch {self.sketch_id}')
-    while timelines:
-      for timeline in timelines:
-        # if the timeline is is a final state, pop it from the list
-        if timeline.status in ['fail', 'ready', 'timeout', 'archived']:
-          timelines.remove(timeline)
-          # if the timeline is ready, run the analyzers
-          if timeline.status == 'ready' and (not timeline.id
-                                             in self._processed_timelines):
-            self._processed_timelines.add(timeline.id)
-            self._RunAnalyzers(timeline.name)
-        else:
-          self.logger.info(f'Waiting for timeline {timeline.name} to be ready')
-      time.sleep(30)
+    while True:
+      if all(tl.status in ['fail', 'ready', 'timeout', 'archived']
+             for tl in timelines):
+        break
+      time.sleep(10)
 
-  def _RunAnalyzers(self, timeline_name: str) -> None:
-    """Runs analyzers on a timeline."""
-    if not self._analyzers:
-      self.logger.info(
-          'No analyzers to run on timeline {0:s}.'.format(timeline_name))
-      return
-
-    timeline = self.sketch.get_timeline(timeline_name=timeline_name)
-    self.logger.info(
-        "Running analyzers {0!s} on timeline {1:s}".format(
-            self._analyzers, timeline_name))
-    try:
-      # By default run_analyzers() ignores analyzers that have already been run
-      results: List[ts_analyzer.AnalyzerResult] = timeline.run_analyzers(
-          analyzer_names=self._analyzers)
-      if not results:
-        self.logger.info(
-            'No new analyzers to run on timeline {0:s}.'.format(timeline_name))
-        return
-      # Get the last result, which is the most recent run of the analyzers.
-      result = results[-1]
-      for analyzer_name, analyzer_status in result.status_dict.items():
-        self.logger.debug(
-            'Analyzer: {0:s} status: {1!s}'.format(
-                analyzer_name, analyzer_status))
-    except ts_error.UnableToRunAnalyzer as exception:
-      self.ModuleError(
-          'Unable to run analyzer: {0!s}'.format(exception), critical=False)
-
-  # pytype: disable=signature-mismatch
-  def Process(self, container: containers.File) -> None:
+  def Process(self, container: containers.File
+              ) -> None:  # pytype: disable=signature-mismatch
     """Executes a Timesketch export.
 
     Args:
       container (containers.File): A container holding a File to import."""
 
     recipe_name = self.state.recipe.get('name', 'no_recipe')
-    rand = uuid.uuid4().hex[-5:]
+
     description = container.name
     if description:
       name = description.rpartition('.')[0]
       name = name if name else description
       name = name.replace(' ', '_').replace('-', '_')
-      timeline_name = f'{recipe_name}_{name}'
+      timeline_name = '{0:s}_{1:s}'.format(
+          recipe_name, name)
     else:
-      timeline_name = f'{recipe_name}'
+      timeline_name = recipe_name
 
-    # Give each timeline a unique name
-    timeline_name = f'{timeline_name}_{rand}'
-    self.logger.info('Uploading {0:s} ...'.format(timeline_name))
+    self.logger.info('Uploading {0:s}...'.format(timeline_name))
 
     with importer.ImportStreamer() as streamer:
       streamer.set_sketch(self.sketch)
       streamer.set_timeline_name(timeline_name)
 
       path = container.path
-      try:
-        streamer.add_file(path)
-      except RuntimeError as exception:
-        self.ModuleError(
-            'Unable to import {0:s}: {1!s}'.format(path, exception),
-            critical=False)
+      streamer.add_file(path)
       if streamer.response and container.description:
         streamer.timeline.description = container.description
+
+    if self.wait_for_timelines:
+      self.logger.info('Waiting for timeline {0:s} to finish processing...'\
+          .format(timeline_name))
+      self._WaitForTimelines()
+
+    for analyzer in self._analyzers:
+      self.logger.info("Running analyzer {0:s} on timeline {1:s}".format(
+          analyzer, timeline_name))
+      results = self.sketch.run_analyzer(
+          analyzer_name=analyzer, timeline_name=timeline_name)
+
+      if not results:
+        self.logger.info('Analyzer [{0:s}] not able to run on {1:s}'.format(
+            analyzer, timeline_name))
+        continue
+
+      # Unknown why, but we get a list of 2 identical result objects
+      results = results[0]
+      session_id = results._session_id # pylint: disable=protected-access
+      if not session_id:
+        self.logger.info(
+            'Analyzer [{0:s}] didn\'t provide any session data'.format(
+                analyzer))
+        continue
+      self.logger.info('Analyzer: {0:s} is running, session ID: {1:d}'.format(
+          analyzer, session_id))
+      self.logger.info(results.status_string)
 
   def GetThreadOnContainerType(self) -> Type[interface.AttributeContainer]:
     return containers.File
@@ -245,21 +206,33 @@ class TimesketchExporter(module.ThreadAwareModule):
     return 5
 
   def PreProcess(self) -> None:
-    pass
+    """Get the sketch, creating it if it doesn't yet exist."""
+    if not self.timesketch_api:
+      message = 'Could not connect to Timesketch server'
+      self.ModuleError(message, critical=True)
+
+    self.sketch = self.state.GetFromCache('timesketch_sketch')
+    if not self.sketch and self.sketch_id:
+      self.logger.info('Using exiting sketch: {0:d}'.format(self.sketch_id))
+      self.sketch = self.timesketch_api.get_sketch(self.sketch_id)
+
+    # Create the sketch if no sketch was stored in the cache.
+    if not self.sketch:
+      self.sketch = self._CreateSketch(incident_id=self.incident_id)
+      self.sketch_id = self.sketch.id
+      self.logger.info('New sketch created: {0:d}'.format(self.sketch_id))
 
   def PostProcess(self) -> None:
     api_root = self.sketch.api.api_root
     host_url = api_root.partition('api/v1')[0]
-
-    if self.wait_for_timelines:
-      self._WaitForTimelines()
-
     sketch_url = '{0:s}sketches/{1:d}/'.format(host_url, self.sketch.id)
     message = 'Your Timesketch URL is: {0:s}'.format(sketch_url)
     self.PublishMessage(message)
 
     report_container = containers.Report(
-        module_name='TimesketchExporter', text=message, text_format='markdown')
+        module_name='TimesketchExporter',
+        text=message,
+        text_format='markdown')
     self.StoreContainer(report_container)
