+++ dftimewolf/lib/processors/turbinia_gcp.py
@@ -1,8 +1,13 @@
+# -*- coding: utf-8 -*-
 """Processes GCP cloud disks using Turbinia."""
 
-from typing import Any, Dict, Optional, TYPE_CHECKING, Type, Union, Set
+import os
+from typing import Dict, List, Optional, TYPE_CHECKING, Any, Type, Union
 
 import magic
+from turbinia import TurbiniaException
+from turbinia import config as turbinia_config  #pylint: disable=unused-import
+from turbinia import evidence
 
 from dftimewolf.lib import module
 from dftimewolf.lib.containers import containers, interface
@@ -17,14 +22,13 @@ class TurbiniaGCPProcessor(TurbiniaProcessorBase, module.ThreadAwareModule):
   """Processes Google Cloud (GCP) disks with Turbinia.
 
   Attributes:
-    request_ids Set[str]: Turbinia requests for jobs being processed.
+    disk_name (str): name of the disk to process.
   """
 
-  def __init__(
-      self,
-      state: "state.DFTimewolfState",
-      name: Optional[str] = None,
-      critical: bool = False) -> None:
+  def __init__(self,
+               state: "state.DFTimewolfState",
+               name: Optional[str]=None,
+               critical: bool=False) -> None:
     """Initializes a Turbinia Google Cloud (GCP) disks processor.
 
     Args:
@@ -34,48 +38,79 @@ class TurbiniaGCPProcessor(TurbiniaProcessorBase, module.ThreadAwareModule):
           the entire recipe to fail if the module encounters an error.
     """
     module.ThreadAwareModule.__init__(self, state, name=name, critical=critical)
-    TurbiniaProcessorBase.__init__(
-        self, state, self.logger, name=name, critical=critical)
-    self.request_ids: Set[str] = set()
+    TurbiniaProcessorBase.__init__(self, self.logger)
 
-  def _BuildContainer(
-      self, path: str, container_name: str
-  ) -> Optional[Union[containers.File, containers.ThreatIntelligence]]:
-    """Builds a container from a path."""
-    container: Optional[Union[containers.File,
-                              containers.ThreatIntelligence]] = None
-    if path.endswith('BinaryExtractorTask.tar.gz'):
-      container = containers.ThreatIntelligence(
-          name=container_name, indicator=None, path=path)
-    elif path.endswith('hashes.json'):
-      container = containers.ThreatIntelligence(
-          name=container_name, indicator=None, path=path)
-    elif path.endswith('.plaso'):
-      container = containers.File(name=container_name, path=path)
-    elif magic.from_file(path, mime=True).startswith('text'):
-      container = containers.File(name=container_name, path=path)
-    else:
-      self.PublishMessage(
-          f'Skipping result of type {magic.from_file(path)} at: {path}')
-
-    return container
-
-  def _CreateTurbiniaRequest(
-      self, request_container: containers.TurbiniaRequest) -> str:
-    """Creates a new Turbinia request.
+  # pylint: disable=arguments-differ
+  def SetUp(self,
+            turbinia_config_file: Union[str, None],
+            project: str,
+            turbinia_recipe: Union[str, None],
+            turbinia_zone: str,
+            sketch_id: int,
+            disk_names: str = '') -> None:
+    """Sets up the object attributes.
 
     Args:
-      request_container (containers.TurbiniaRequest): Turbinia request.
-
-    Returns:
-      str: Turbinia request identifier.
+      turbinia_config_file (str): Full path to the Turbinia config file to use.
+      disk_names (str): names of the disks to process.
+      project (str): name of the GCP project containing the disk to process.
+      turbinia_recipe (str): Turbinia recipe name.
+      turbinia_zone (str): GCP zone in which the Turbinia server is running.
+      sketch_id (int): The Timesketch sketch ID.
     """
-    evidence = {
-        'type': 'GoogleCloudDisk',
-        'disk_name': request_container.evidence_name,
-        'project': self.project,
-        'zone': self.turbinia_zone
-    }
+    self.turbinia_config_file = turbinia_config_file
+
+    if disk_names:
+      for disk in disk_names.strip().split(','):
+        if not disk:
+          continue
+        self.StoreContainer(containers.GCEDisk(
+            name=disk,
+            project=project))
+
+    try:
+      self.TurbiniaSetUp(project, turbinia_recipe, turbinia_zone, sketch_id)
+    except TurbiniaException as exception:
+      self.ModuleError(str(exception), critical=True)
+      return
+
+  def PreProcess(self) -> None:
+    """Ensure ForensicsVM containers from previous modules are processed.
+
+    Before the addition of containers.GCEDiskEvidence, containers.ForensicsVM
+    was used to track disks needing processing by Turbinia via this module. Here
+    we grab those containers and track the disks for processing by this module,
+    for any modules that aren't using the new container yet.
+    """
+    vm_containers = self.GetContainers(containers.ForensicsVM)
+    for container in vm_containers:
+      if container.evidence_disk and container.evidence_disk.name:
+        self.StoreContainer(
+            containers.GCEDisk(
+                name=container.evidence_disk.name,
+                project=self.project))
+    self.state.DedupeContainers(containers.GCEDisk)
+
+  # pylint: disable=arguments-renamed
+  def Process(self, disk_container: containers.GCEDisk) -> None:
+    """Process a GCE Disk with Turbinia."""
+    if disk_container.project != self.project:
+      self.logger.info(f'Found disk "{disk_container.name}" but skipping as it '
+          f'is in a different project "{disk_container.project}".')
+      return
+
+    log_file_path = os.path.join(
+        self._output_path, f'{disk_container.name}-turbinia.log')
+
+    self.logger.info(f'Turbinia log file: {log_file_path}')
+    self.logger.info(
+        f'Using disk {disk_container.name} from previous collector')
+
+    evidence_ = evidence.GoogleCloudDisk(
+        disk_name=disk_container.name,
+        project=self.project,
+        zone=self.turbinia_zone)
+
     threat_intel_indicators = None
     threatintel = self.GetContainers(containers.ThreatIntelligence)
     if threatintel:
@@ -86,153 +121,70 @@ class TurbiniaGCPProcessor(TurbiniaProcessorBase, module.ThreadAwareModule):
     yara_rules = None
     yara_containers = self.GetContainers(containers.YaraRule)
     if yara_containers:
-      self.logger.info(
-          f'Sending {len(yara_containers)} Yara rules to Turbinia '
+      self.logger.info(f'Sending {len(yara_containers)} Yara rules to Turbinia '
           'Plaso worker...')
       yara_rules = [rule.rule_text for rule in yara_containers]
 
-    request_id = self.TurbiniaStart(
-        evidence, threat_intel_indicators, yara_rules)
-    if not request_id:
-      self.ModuleError('Turbinia request failed', critical=True)
+    try:
+      request_id = self.TurbiniaStart(
+          evidence_, threat_intel_indicators, yara_rules)
+      self.PublishMessage(f'Turbinia request ID: {request_id}')
+      task_data, report = self.TurbiniaWait(request_id)
+    except TurbiniaException as exception:
+      self.ModuleError(str(exception), critical=True)
 
-    return request_id
+    self.StoreContainer(containers.Report(
+        module_name='TurbiniaProcessor', text=report, text_format='markdown'))
 
-  # pylint: disable=arguments-differ
-  def SetUp(
-      self,
-      project: str,
-      turbinia_auth: bool,
-      turbinia_recipe: Union[str, None],
-      turbinia_zone: str,
-      turbinia_api: str,
-      incident_id: str,
-      sketch_id: int,
-      request_ids: str = '',
-      disk_names: str = '') -> None:
-    """Sets up the object attributes.
+    local_paths, gs_paths = self._DeterminePaths(task_data)
 
-    Args:
-      disk_names (str): names of the disks to process.
-      project (str): name of the GCP project containing the disk to process.
-      turbinia_auth (bool): Turbinia auth flag.
-      turbinia_api (str): Turbinia API endpoint.
-      turbinia_recipe (str): Turbinia recipe name.
-      turbinia_zone (str): GCP zone in which the Turbinia server is running.
-      incident_id (str): The incident ID.
-      sketch_id (int): The sketch ID.
-      request_ids (str): Turbinia requests for jobs being processed.
-      disk_names (str): Names of the disks to process.
-    """
-
-    if (disk_names and request_ids):
+    if not local_paths and not gs_paths:
       self.ModuleError(
-          'One of disk_names or request_ids can be specified, but not both.',
-          critical=True)
-      return
+          'No interesting files found in Turbinia output.', critical=True)
 
-    if request_ids:
-      self.request_ids = {
-          request_ids.strip()
-          for request_ids in request_ids.split(',')
-          if request_ids.strip()
-      }
-      for request_id in self.request_ids:
-        self.StoreContainer(
-            containers.TurbiniaRequest(project=project, request_id=request_id))
+    timeline_label = f'{self.project}-{disk_container.name}'
+    # Any local files that exist we can add immediately to the output
+    all_local_paths = [
+        (timeline_label, p) for p in local_paths if os.path.exists(p)]
 
-    if disk_names:
-      for disk in disk_names.strip().split(','):
-        if not disk:
-          continue
-        self.StoreContainer(
-            containers.TurbiniaRequest(project=project, evidence_name=disk))
+    try:
+      downloaded_gs_paths = self._DownloadFilesFromGCS(timeline_label, gs_paths)
+    except TurbiniaException as exception:
+      # Don't add a critical error for now, until we start raising errors
+      # instead of returning manually each
+      self.ModuleError(str(exception), critical=False)
 
-    self.TurbiniaSetUp(
-        project, turbinia_auth, turbinia_recipe, turbinia_zone, turbinia_api,
-        incident_id, sketch_id)
+    all_local_paths.extend(downloaded_gs_paths)
+    self.logger.info(f'Collected {len(all_local_paths)} results')
 
-  def PreProcess(self) -> None:
-    """Ensures containers from previous modules are processed.
+    if not all_local_paths:
+      self.ModuleError('No interesting files could be found.', critical=True)
 
-    GCEDisk containers from preivous modules will be deduplicated, and
-    TurbiniaRequest containers will be created for each GCEDisk container.
-    This is necessary because TurbiniaRequest containers are used to track
-    Turbinia jobs (e.g. to support resuming a recipe after a failure).
-    """
-    vm_containers = self.GetContainers(containers.ForensicsVM)
-    for container in vm_containers:
-      if container.evidence_disk and container.evidence_disk.name:
-        self.StoreContainer(
-            containers.GCEDisk(
-                name=container.evidence_disk.name, project=self.project))
-
-    self.state.DedupeContainers(containers.GCEDisk)
-
-    disk_containers = self.GetContainers(containers.GCEDisk)
-    turb_containers = self.GetContainers(containers.TurbiniaRequest)
-    if not disk_containers and not turb_containers:
-      self.ModuleError(
-          'No disk names or request IDs specified, and there are no valid '
-          'containers to process from previous modules. '
-          'Please specify disk names or request IDs.',
-          critical=True)
-      return
-    for disk_container in disk_containers:
-      self.StoreContainer(
-          containers.TurbiniaRequest(
-              project=disk_container.project,
-              evidence_name=disk_container.name))
-
-  # pylint: disable=arguments-renamed
-  def Process(self, request_container: containers.TurbiniaRequest) -> None:
-    """Process a GCE Disk with Turbinia."""
-    request_id = ''
-    task: Dict[str, Any] = {}
-    report = ''
-
-    if request_container.project != self.project:
-      self.logger.info(
-          f'Found disk "{request_container.evidence_name}" but skipping as it '
-          f'is in a different project "{request_container.project}".')
-      return
-
-    if request_container.request_id:
-      # We have a request ID, so we can skip creating a new Turbinia request.
-      request_id = request_container.request_id
-    else:
-      # We don't have a request ID, so we need to create a new Turbinia request.
-      request_id = self._CreateTurbiniaRequest(request_container)
-
-    self.PublishMessage(f'Turbinia request identifier: {request_id}')
-
-    for task, path in self.TurbiniaWait(request_id):
-      task_id = task.get('id')
-      task_name = task.get('name')
-      container_name = f'{self.project}-{task_name}-{task_id}'
-      self.PublishMessage(f'New output file {path} found for task {task_id}')
-      local_path = self.DownloadFilesFromAPI(task, path)
-      if not local_path:
-        self.logger.warning(
-            f'No interesting output files could be found for task {task_id}')
+    container: Union[containers.File, containers.ThreatIntelligence]
+    for description, path in all_local_paths:
+      if path.endswith('BinaryExtractorTask.tar.gz'):
+        self.PublishMessage(f'Found BinaryExtractorTask result: {path}')
+        container = containers.ThreatIntelligence(
+            name='BinaryExtractorResults', indicator=None, path=path)
+      elif path.endswith('hashes.json'):
+        self.PublishMessage(f'Found hashes.json: {path}')
+        container = containers.ThreatIntelligence(
+            name='ImageExportHashes', indicator=None, path=path)
+      elif path.endswith('.plaso'):
+        self.PublishMessage(f'Found plaso result: {path}')
+        container = containers.File(name=description, path=path)
+      elif magic.from_file(path, mime=True).startswith('text'):
+        self.PublishMessage(f'Found result: {path}')
+        container = containers.File(name=description, path=path)
+      else:
+        self.PublishMessage(
+            f'Skipping result of type {magic.from_file(path)} at: {path}')
         continue
-      container = self._BuildContainer(local_path, container_name)
-      if container:
-        self.PublishMessage(f'Streaming container {container.name}')
-        self.StreamContainer(container)
+      self.StoreContainer(container)
+  # pylint: enable=arguments-renamed
 
-    # Generate a Turbinia report and store it in the state.
-    report = self.TurbiniaFinishReport(request_id)
-    self.StoreContainer(
-        containers.Report(
-            module_name='TurbiniaProcessor',
-            text=report,
-            text_format='markdown'))
-    self.PublishMessage(report)
-
-  @staticmethod
-  def GetThreadOnContainerType() -> Type[interface.AttributeContainer]:
-    return containers.TurbiniaRequest
+  def GetThreadOnContainerType(self) -> Type[interface.AttributeContainer]:
+    return containers.GCEDisk
 
   def GetThreadPoolSize(self) -> int:
     return self.parallel_count
