+++ dftimewolf/lib/state.py
@@ -23,7 +23,6 @@ from dftimewolf.lib.module import ThreadAwareModule, BaseModule
 
 if TYPE_CHECKING:
   from dftimewolf.lib import module as dftw_module
-
 T = TypeVar("T", bound="interface.AttributeContainer")  # pylint: disable=invalid-name,line-too-long
 
 # TODO(tomchop): Consider changing this to `dftimewolf.state` if we ever need
@@ -33,7 +32,6 @@ logger = logging.getLogger('dftimewolf')
 NEW_ISSUE_URL = 'https://github.com/log2timeline/dftimewolf/issues/new'
 
 TELEMETRY = telemetry
-
 
 class DFTimewolfState(object):
   """The main State class.
@@ -53,7 +51,8 @@ class DFTimewolfState(object):
     telemetry_store: store for statistics generated by modules.
   """
 
-  def __init__(self, config: Type[Config]) -> None:
+  def __init__(self,
+               config: Type[Config]) -> None:
     """Initializes a state."""
     super(DFTimewolfState, self).__init__()
     self.command_line_options = {}  # type: Dict[str, Any]
@@ -63,8 +62,8 @@ class DFTimewolfState(object):
     self._threading_event_per_module = {}  # type: Dict[str, threading.Event]
     self.config = config
     self.errors = []  # type: List[DFTimewolfError]
-    self.global_errors = []  # type: List[DFTimewolfError]
-    self.recipe = {}  # type: Dict[str, Any]
+    self.global_errors = [] # type: List[DFTimewolfError]
+    self.recipe = {} # type: Dict[str, Any]
     self.store = {}  # type: Dict[str, List[interface.AttributeContainer]]
     self.streaming_callbacks = {}  # type: Dict[Type[interface.AttributeContainer], List[Callable[[Any], Any]]]  # pylint: disable=line-too-long
     self._abort_execution = False
@@ -79,7 +78,7 @@ class DFTimewolfState(object):
     """
     threads = []
     for module_definition in self.recipe['modules']:
-      thread_args = (module_definition,)
+      thread_args = (module_definition, )
       thread = threading.Thread(target=callback, args=thread_args)
       threads.append(thread)
       thread.start()
@@ -103,18 +102,12 @@ class DFTimewolfState(object):
     """
     for module in self.recipe['modules'] + self.recipe.get('preflights', []):
       name = module['name']
-      if name.endswith('Legacy'):
-        msg = f'Skipping legacy module {name}, rename module to use.'
-        logger.warning(msg)
-        continue
       if name not in module_locations:
-        msg = (
-            f'In {self.recipe["name"]}: module {name} cannot be found. '
-            'It may not have been declared.')
+        msg = (f'In {self.recipe["name"]}: module {name} cannot be found. '
+               'It may not have been declared.')
         raise errors.RecipeParseError(msg)
-      logger.debug(
-          'Loading module {0:s} from {1:s}'.format(
-              name, module_locations[name]))
+      logger.debug('Loading module {0:s} from {1:s}'.format(
+          name, module_locations[name]))
 
       location = module_locations[name]
       try:
@@ -123,8 +116,9 @@ class DFTimewolfState(object):
         msg = f'Cannot find Python module for {name} ({location}): {exception}'
         raise errors.RecipeParseError(msg)
 
-  def LoadRecipe(
-      self, recipe: Dict[str, Any], module_locations: Dict[str, str]) -> None:
+  def LoadRecipe(self,
+                 recipe: Dict[str, Any],
+                 module_locations: Dict[str, str]) -> None:
     """Populates the internal module pool with modules declared in a recipe.
 
     Args:
@@ -141,14 +135,14 @@ class DFTimewolfState(object):
     for module_definition in module_definitions + preflight_definitions:
       # Combine CLI args with args from the recipe description
       module_name = module_definition['name']
+      module_class = modules_manager.ModulesManager.GetModuleByName(module_name)
+
       runtime_name = module_definition.get('runtime_name')
       if not runtime_name:
         runtime_name = module_name
-      module_class = modules_manager.ModulesManager.GetModuleByName(module_name)
-      if module_class:
-        # pytype: disable=wrong-arg-types
-        self._module_pool[runtime_name] = module_class(self, name=runtime_name)
-        # pytype: enable=wrong-arg-types
+      # pytype: disable=wrong-arg-types
+      self._module_pool[runtime_name] = module_class(self, name=runtime_name)
+      # pytype: enable=wrong-arg-types
 
   def FormatExecutionPlan(self) -> str:
     """Formats execution plan.
@@ -202,7 +196,7 @@ class DFTimewolfState(object):
     with self._state_lock:
       self._cache[name] = value
 
-  def GetFromCache(self, name: str, default_value: Any = None) -> Any:
+  def GetFromCache(self, name: str, default_value: Any=None) -> Any:
     """Thread-safe method to get data from the state's cache.
 
     Args:
@@ -219,10 +213,9 @@ class DFTimewolfState(object):
     with self._state_lock:
       return self._cache.get(name, default_value)
 
-  def StoreContainer(
-      self,
-      container: "interface.AttributeContainer",
-      source_module: str = "") -> None:
+  def StoreContainer(self,
+                     container: "interface.AttributeContainer",
+                     source_module: str = "") -> None:
     """Thread-safe method to store data in the state's store.
 
     Args:
@@ -230,11 +223,12 @@ class DFTimewolfState(object):
       source_module: the originating module.
     """
     with self._state_lock:
-      container.SetMetadata(interface.METADATA_KEY_SOURCE_MODULE, source_module)
+      container.SetMetadata(
+          interface.METADATA_KEY_SOURCE_MODULE, source_module)
       self.store.setdefault(container.CONTAINER_TYPE, []).append(container)
 
   def LogTelemetry(
-      self, telemetry_entry: telemetry.TelemetryCollection) -> None:
+    self, telemetry_entry: telemetry.TelemetryCollection) -> None:
     """Method to store telemetry in the state's telemetry store.
 
     Args:
@@ -242,14 +236,14 @@ class DFTimewolfState(object):
     """
     for key, value in telemetry_entry.telemetry.items():
       TELEMETRY.LogTelemetry(
-          key, value, telemetry_entry.module_name, telemetry_entry.recipe)
+        key, value, telemetry_entry.module_name, telemetry_entry.recipe
+      )
 
-  def GetContainers(
-      self,
-      container_class: Type[T],
-      pop: bool = False,
-      metadata_filter_key: Optional[str] = None,
-      metadata_filter_value: Optional[Any] = None) -> Sequence[T]:
+  def GetContainers(self,
+                    container_class: Type[T],
+                    pop: bool=False,
+                    metadata_filter_key: Optional[str]=None,
+                    metadata_filter_value: Optional[Any]=None) -> Sequence[T]:
     """Thread-safe method to retrieve data from the state's store.
 
     Args:
@@ -332,18 +326,16 @@ class DFTimewolfState(object):
       # We're catching any exception that is not a DFTimewolfError, so we want
       # to generate an error for further reporting.
       error = errors.DFTimewolfError(
-          message=msg,
-          name='dftimewolf',
-          stacktrace=traceback.format_exc(),
-          critical=True,
-          unexpected=True)
+          message=msg, name='dftimewolf', stacktrace=traceback.format_exc(),
+          critical=True, unexpected=True)
       self.AddError(error)
 
     self._threading_event_per_module[runtime_name] = threading.Event()
     self.CleanUp()
 
-  def _RunModuleSetUp(
-      self, module: BaseModule, **new_args: Dict[str, object]) -> None:
+  def _RunModuleSetUp(self,
+                     module: BaseModule,
+                     **new_args: Dict[str, object]) -> None:
     """Runs SetUp of a single module.
 
     Designed to be wrapped by an output handling subclass.
@@ -363,7 +355,8 @@ class DFTimewolfState(object):
     module.Process()
 
   def _RunModuleProcessThreaded(
-      self, module: ThreadAwareModule) -> List[Future]:  # type: ignore
+      self, module: ThreadAwareModule
+    ) -> List[Future]: # type: ignore
     """Runs Process of a single ThreadAwareModule module.
 
     Designed to be wrapped by an output handling subclass.
@@ -377,8 +370,8 @@ class DFTimewolfState(object):
     """
     cont_count = len(self.GetContainers(module.GetThreadOnContainerType()))
     logger.info(
-        f'Running {cont_count} threads, max {module.GetThreadPoolSize()} '
-        f'simultaneous for module {module.name}')
+       f'Running {cont_count} threads, max {module.GetThreadPoolSize()} '
+       f'simultaneous for module {module.name}')
 
     futures = []
 
@@ -390,7 +383,8 @@ class DFTimewolfState(object):
             f'Launching {module.name}.Process thread with {str(c)} from '
             f'{c.metadata.get(interface.METADATA_KEY_SOURCE_MODULE, "Unknown")}'
         )
-        futures.append(executor.submit(module.Process, c))
+        futures.append(
+            executor.submit(module.Process, c))
 
     return futures
 
@@ -424,8 +418,7 @@ class DFTimewolfState(object):
       runtime_name: runtime name of the module."""
     for fut in futures:
       if fut.exception():
-        raise fut.exception()  # type: ignore
-
+        raise fut.exception() # type: ignore
   # pylint: disable=unused-argument
 
   def SetupModules(self) -> None:
@@ -486,11 +479,8 @@ class DFTimewolfState(object):
       # We're catching any exception that is not a DFTimewolfError, so we want
       # to generate an error for further reporting.
       error = errors.DFTimewolfError(
-          message=msg,
-          name='dftimewolf',
-          stacktrace=traceback.format_exc(),
-          critical=True,
-          unexpected=True)
+          message=msg, name='dftimewolf', stacktrace=traceback.format_exc(),
+          critical=True, unexpected=True)
       self.AddError(error)
 
     logger.info('Module {0:s} finished execution'.format(runtime_name))
@@ -525,7 +515,7 @@ class DFTimewolfState(object):
       finally:
         self.CheckErrors(is_global=True)
 
-  def InstantiateModule(self, module_name: str) -> Optional["BaseModule"]:
+  def InstantiateModule(self, module_name: str) -> "BaseModule":
     """Instantiates an arbitrary dfTimewolf module.
 
     Args:
@@ -533,22 +523,22 @@ class DFTimewolfState(object):
 
     Returns:
       BaseModule: An instance of a dftimewolf Module, which is a subclass of
-          BaseModule, or None if the module could not be found.
+          BaseModule.
     """
-    module_class: Optional[Type["BaseModule"]]
+    module_class: Type["BaseModule"]
     module_class = modules_manager.ModulesManager.GetModuleByName(module_name)
     # pytype: disable=wrong-arg-types
-    if module_class:
-      return module_class(self)
+    return module_class(self)
     # pytype: enable=wrong-arg-types
-    return None
 
   def RunModules(self) -> None:
     """Performs the actual processing for each module in the module pool."""
     self._InvokeModulesInThreads(self._RunModuleThread)
 
   def RegisterStreamingCallback(
-      self, target: Callable[[T], Any], container_type: Type[T]) -> None:
+      self,
+      target: Callable[[T], Any],
+      container_type: Type[T]) -> None:
     """Registers a callback for a type of container.
 
     The function to be registered should a single parameter of type
@@ -563,10 +553,9 @@ class DFTimewolfState(object):
       self.streaming_callbacks[container_type] = []
     self.streaming_callbacks[container_type].append(target)
 
-  def StreamContainer(
-      self,
-      container: "interface.AttributeContainer",
-      source_module: str = "") -> None:
+  def StreamContainer(self,
+                      container: "interface.AttributeContainer",
+                      source_module: str = "") -> None:
     """Streams a container to the callbacks that are registered to handle it.
 
     Args:
@@ -598,7 +587,7 @@ class DFTimewolfState(object):
     self.global_errors.extend(self.errors)
     self.errors = []
 
-  def CheckErrors(self, is_global: bool = False) -> None:
+  def CheckErrors(self, is_global: bool=False) -> None:
     """Checks for errors and exits if any of them are critical.
 
     Args:
@@ -615,9 +604,8 @@ class DFTimewolfState(object):
       logger.error('dfTimewolf encountered one or more errors:')
 
     for index, error in enumerate(error_objects):
-      logger.error(
-          '{0:d}: error from {1:s}: {2:s}'.format(
-              index + 1, error.name, error.message))
+      logger.error('{0:d}: error from {1:s}: {2:s}'.format(
+          index+1, error.name, error.message))
       if error.stacktrace:
         for line in error.stacktrace.split('\n'):
           logger.error(line)
@@ -632,8 +620,10 @@ class DFTimewolfState(object):
     if critical_errors:
       raise errors.CriticalError('Critical error found. Aborting.')
 
-  def PublishMessage(
-      self, source: str, message: str, is_error: bool = False) -> None:
+  def PublishMessage(self,
+                     source: str,
+                     message: str,
+                     is_error: bool = False) -> None:
     """Receives a message for publishing.
 
     The base class does nothing with this (as the method in module also logs the
@@ -645,16 +635,20 @@ class DFTimewolfState(object):
       is_error: True if the message is an error message, False otherwise.
     """
 
-  def ProgressUpdate(
-      self, module_name: str, steps_taken: int, steps_expected: int) -> None:
+  def ProgressUpdate(self,
+                     module_name: str,
+                     steps_taken: int,
+                     steps_expected: int) -> None:
     """Currently unsupported when no UI is in use."""
     if not self._progress_warning_shown:
       self._progress_warning_shown = True
       logger.debug('ProgressUpdate called in unsupported display mode.')
 
-  def ThreadProgressUpdate(
-      self, module_name: str, thread_id: str, steps_taken: int,
-      steps_expected: int) -> None:
+  def ThreadProgressUpdate(self,
+                           module_name: str,
+                           thread_id: str,
+                           steps_taken: int,
+                           steps_expected: int) -> None:
     """Currently unsupported when no UI is in use."""
     if not self._progress_warning_shown:
       self._progress_warning_shown = True
@@ -665,15 +659,17 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
   """The main state class, extended to wrap methods with updates to a
   CursesDisplayManager object."""
 
-  def __init__(
-      self, config: Type[Config], cursesdm: cdm.CursesDisplayManager) -> None:
+  def __init__(self,
+               config: Type[Config],
+               cursesdm: cdm.CursesDisplayManager) -> None:
     """Initializes a state."""
     super(DFTimewolfStateWithCDM, self).__init__(config)
     self.cursesdm = cursesdm
     self.stdout_log = False
 
-  def LoadRecipe(
-      self, recipe: Dict[str, Any], module_locations: Dict[str, str]) -> None:
+  def LoadRecipe(self,
+                 recipe: Dict[str, Any],
+                 module_locations: Dict[str, str]) -> None:
     """Populates the internal module pool with modules declared in a recipe.
 
     Args:
@@ -689,17 +685,18 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
 
     self.cursesdm.SetRecipe(self.recipe['name'])
     for module_definition in preflight_definitions:
-      self.cursesdm.EnqueuePreflight(
-          module_definition['name'], module_definition.get('wants', []),
-          module_definition.get('runtime_name'))
+      self.cursesdm.EnqueuePreflight(module_definition['name'],
+                                     module_definition.get('wants', []),
+                                     module_definition.get('runtime_name'))
     for module_definition in module_definitions:
-      self.cursesdm.EnqueueModule(
-          module_definition['name'], module_definition.get('wants', []),
-          module_definition.get('runtime_name'))
+      self.cursesdm.EnqueueModule(module_definition['name'],
+                                  module_definition.get('wants', []),
+                                  module_definition.get('runtime_name'))
     self.cursesdm.Draw()
 
-  def _RunModuleSetUp(
-      self, module: BaseModule, **new_args: Dict[str, object]) -> None:
+  def _RunModuleSetUp(self,
+                     module: BaseModule,
+                     **new_args: Dict[str, object]) -> None:
     """Runs SetUp of a single module.
 
     Args:
@@ -719,7 +716,8 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
     self.cursesdm.UpdateModuleStatus(module.name, cdm.Status.COMPLETED)
 
   def _RunModuleProcessThreaded(
-      self, module: ThreadAwareModule) -> List[Future]:  # type: ignore
+      self, module: ThreadAwareModule
+    ) -> List[Future]: # type: ignore
     """Runs Process of a single ThreadAwareModule module.
 
     Args:
@@ -731,8 +729,8 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
     """
     cont_count = len(self.GetContainers(module.GetThreadOnContainerType()))
     logger.info(
-        f'Running {cont_count} threads, max {module.GetThreadPoolSize()} '
-        f'simultaneous for module {module.name}')
+       f'Running {cont_count} threads, max {module.GetThreadPoolSize()} '
+       f'simultaneous for module {module.name}')
 
     self.cursesdm.SetThreadedModuleContainerCount(module.name, cont_count)
     self.cursesdm.UpdateModuleStatus(module.name, cdm.Status.PROCESSING)
@@ -744,7 +742,8 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
       pop = not module.KeepThreadedContainersInState()
       for c in self.GetContainers(module.GetThreadOnContainerType(), pop):
         futures.append(
-            executor.submit(self._WrapThreads, module.Process, c, module.name))
+            executor.submit(
+                self._WrapThreads, module.Process, c, module.name))
 
     return futures
 
@@ -780,9 +779,10 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
         self.cursesdm.SetError(runtime_name, str(fut.exception()))
         raise fut.exception()  # type: ignore
 
-  def _WrapThreads(
-      self, process: Callable[[AttributeContainer], None],
-      container: AttributeContainer, module_name: str) -> None:
+  def _WrapThreads(self,
+                  process: Callable[[AttributeContainer], None],
+                  container: AttributeContainer,
+                  module_name: str) -> None:
     """Wraps a ThreadPoolExecutor call to module.process with the
     CursesDisplayManager status update methods.
 
@@ -811,8 +811,10 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
     name = error.name if error.name else 'no_module_name'
     self.cursesdm.SetError(name, error.message)
 
-  def PublishMessage(
-      self, source: str, message: str, is_error: bool = False) -> None:
+  def PublishMessage(self,
+                     source: str,
+                     message: str,
+                     is_error: bool = False) -> None:
     """Receives a message for publishing to the list of messages.
 
     Args:
@@ -821,8 +823,10 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
       is_error: True if the message is an error message, False otherwise."""
     self.cursesdm.EnqueueMessage(source, message, is_error)
 
-  def ProgressUpdate(
-      self, module_name: str, steps_taken: int, steps_expected: int) -> None:
+  def ProgressUpdate(self,
+                     module_name: str,
+                     steps_taken: int,
+                     steps_expected: int) -> None:
     """Set the current completion status of a module.
 
     Args:
@@ -830,11 +834,14 @@ class DFTimewolfStateWithCDM(DFTimewolfState):
       steps_taken: The number of steps taken so far.
       steps_expected: The number of total steps expected for completion.
     """
-    self.cursesdm.SetModuleProgress(module_name, steps_taken, steps_expected)
+    self.cursesdm.SetModuleProgress(
+        module_name, steps_taken, steps_expected)
 
-  def ThreadProgressUpdate(
-      self, module_name: str, thread_id: str, steps_taken: int,
-      steps_expected: int) -> None:
+  def ThreadProgressUpdate(self,
+                           module_name: str,
+                           thread_id: str,
+                           steps_taken: int,
+                           steps_expected: int) -> None:
     """Set the current completion status of a module thread.
 
     Args:
