+++ tests/lib/exporters/timesketch.py
@@ -94,19 +94,15 @@ class TimesketchExporterTest(unittest.TestCase):
     self.assertEqual(report.text_format, 'markdown')
 
   # pylint: disable=invalid-name
-  @mock.patch('dftimewolf.lib.exporters.timesketch.TimesketchExporter._RunAnalyzers')  # pylint: disable=line-too-long'')
   @mock.patch('time.sleep')
   @mock.patch('timesketch_import_client.importer.ImportStreamer')
   @mock.patch('dftimewolf.lib.timesketch_utils.GetApiClient')
-  def testWaitForTimeline(self,
-      mock_GetApiClient,
-      unused_streamer,
-      unused_sleep,
-      mock_RunAnalyzers):
+  def testWaitForTimeline(
+        self, mock_GetApiClient, unused_streamer, unused_sleep):
     """Tests the SetUp function."""
     mock_sketch = mock.Mock(id=1234, my_acl=['write'])
     mock_sketch.api.api_root = 'timesketch.com/api/v1'
-    mock_RunAnalyzers.return_value = None
+
     mock_api_client = mock.Mock()
     mock_api_client.get_sketch.return_value = mock_sketch
     mock_api_client.create_sketch.return_value = mock_sketch
