+++ tests/lib/processors/turbinia_artifact.py
@@ -15,33 +15,26 @@ current_dir = os.path.dirname(os.path.realpath(__file__))
 os.environ['TURBINIA_CONFIG_PATH'] = os.path.join(current_dir, 'test_data')
 # pylint: disable=wrong-import-position
 from dftimewolf.lib.containers import containers
+from dftimewolf.lib.processors import turbinia_artifact
+
 from dftimewolf import config
 
-HAS_TURBINIA = False
+# Manually set TURBINIA_PROJECT to the value we expect.
+# pylint: disable=wrong-import-position, wrong-import-order
+from turbinia import config as turbinia_config
+from turbinia import message as turbinia_message
 
-try:
-  from dftimewolf.lib.processors import turbinia_artifact_legacy
-  from turbinia import config as turbinia_config
-  from turbinia import message as turbinia_message
-  # Manually set TURBINIA_PROJECT to the value we expect.
-  turbinia_config.TURBINIA_PROJECT = 'turbinia-project'
-  HAS_TURBINIA = True
-except ImportError:
-  pass
-
+turbinia_config.TURBINIA_PROJECT = 'turbinia-project'
 YARA_RULE = """rule dummy { condition: false }"""
 
 
-@unittest.skipIf(not HAS_TURBINIA, 'Missing Turbinia dependency.')
 class TurbiniaArtifactProcessorTest(unittest.TestCase):
   """Tests for the Turbinia processor."""
 
   def testInitialization(self):
     """Tests that the processor can be initialized."""
     test_state = state.DFTimewolfState(config.Config)
-    # pylint: disable=line-too-long
-    turbinia_processor = turbinia_artifact_legacy.TurbiniaArtifactProcessorLegacy(
-        test_state)
+    turbinia_processor = turbinia_artifact.TurbiniaArtifactProcessor(test_state)
     self.assertIsNotNone(turbinia_processor)
 
   @mock.patch('turbinia.client.get_turbinia_client')
@@ -49,9 +42,7 @@ class TurbiniaArtifactProcessorTest(unittest.TestCase):
   def testSetup(self, _mock_TurbiniaClient):
     """Tests that the processor is set up correctly."""
     test_state = state.DFTimewolfState(config.Config)
-    # pylint: disable=line-too-long
-    turbinia_processor = turbinia_artifact_legacy.TurbiniaArtifactProcessorLegacy(
-        test_state)
+    turbinia_processor = turbinia_artifact.TurbiniaArtifactProcessor(test_state)
     turbinia_processor.SetUp(
         turbinia_config_file=None,
         project='turbinia-project',
@@ -60,8 +51,7 @@ class TurbiniaArtifactProcessorTest(unittest.TestCase):
         output_directory='/tmp/outputdir',
         sketch_id=123)
     # pylint: disable=line-too-long
-    turbinia_processor.client.create_request.return_value = turbinia_message.TurbiniaRequest(
-    )
+    turbinia_processor.client.create_request.return_value = turbinia_message.TurbiniaRequest()
     self.assertEqual(turbinia_processor.project, 'turbinia-project')
     self.assertEqual(turbinia_processor.turbinia_recipe, None)
     self.assertEqual(turbinia_processor.turbinia_zone, 'europe-west1')
@@ -80,9 +70,7 @@ class TurbiniaArtifactProcessorTest(unittest.TestCase):
         received from the state.
     """
     test_state = state.DFTimewolfState(config.Config)
-    # pylint: disable=line-too-long
-    turbinia_processor = turbinia_artifact_legacy.TurbiniaArtifactProcessorLegacy(
-        test_state)
+    turbinia_processor = turbinia_artifact.TurbiniaArtifactProcessor(test_state)
     turbinia_processor.StoreContainer(
         containers.RemoteFSPath(hostname='remotehost', path='/tmp/file.ext'))
     turbinia_processor.SetUp(
@@ -93,8 +81,7 @@ class TurbiniaArtifactProcessorTest(unittest.TestCase):
         output_directory='/tmp/outputdir',
         sketch_id=123)
     # pylint: disable=line-too-long
-    turbinia_processor.client.create_request.return_value = turbinia_message.TurbiniaRequest(
-    )
+    turbinia_processor.client.create_request.return_value = turbinia_message.TurbiniaRequest()
     turbinia_processor.client.get_task_data.return_value = [{
         'saved_paths': [
             '/fake/data.plaso',
